+++ src/main/java/org/example/tablenow/domain/reservation/message/customer/ReminderRegisterConsumer.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Component;
 
 import java.time.ZoneOffset;
 
-import static org.example.tablenow.global.rabbitmq.constant.RabbitConstant.RESERVATION_REMINDER_REGISTER_QUEUE;
+import static org.example.tablenow.global.constant.RabbitConstant.RESERVATION_REMINDER_REGISTER_QUEUE;
 
 @Slf4j
 @Component
