+++ src/main/java/org/example/tablenow/domain/reservation/message/producer/ReminderRegisterProducer.java
@@ -6,7 +6,7 @@ import org.example.tablenow.domain.reservation.message.dto.ReminderMessage;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.stereotype.Component;
 
-import static org.example.tablenow.global.rabbitmq.constant.RabbitConstant.*;
+import static org.example.tablenow.global.constant.RabbitConstant.*;
 
 @Slf4j
 @Component
